:sparkles: api após mudança no projeto

+++ api/src/main/java/com/api/repositories/EnderecoInterface.java
@@ -3,5 +3,5 @@ package com.api.repositories;
 import com.api.models.EnderecoModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EnderecoInterface extends JpaRepository<EnderecoModel,Long> {
+public interface EnderecoInterface extends JpaRepository<EnderecoModel, Integer> {
 }
